LLM/Zuhayr/nlp_chatbot_ollama/requirements.txt
langchain
langchain-ollama
langchain-chroma
pandas
sqlalchemy
pyodbc 
python-dotenv